+++ example/pages/both/lazy-image/components/Critical.vue
@@ -1,0 +1,37 @@
+<template>
+  <organism-preview-container>
+    <template #default>
+      <lazy-image v-bind="criticalImage" />
+    </template>
+    <template #title>
+      <p>Critical<br>LazyImage - SRCSET<br>Image Switch by >768px</p>
+    </template>
+  </organism-preview-container>
+</template>
+
+<script>
+import LazyImage from 'nuxt-speedkit-components/LazyImage'
+import OrganismPreviewContainer from '@/components/organisms/PreviewContainer'
+export default {
+  components: {
+    OrganismPreviewContainer,
+    LazyImage
+  },
+  data () {
+    const criticalImage = require('@/assets/img/critical-2400.jpg?resize&sizes[]=480,sizes[]=768,sizes[]=960,sizes[]=1080,sizes[]=1200,sizes[]=1536,sizes[]=2160,sizes[]=2400&placeholder&format=webp')
+    const criticalPlaceholder = require('@/assets/img/critical-2400.jpg?sqip')
+
+    return {
+      criticalImage: {
+        caption: null,
+        alt: 'Alt Text',
+        title: 'Title Text',
+        placeholder: (({ src, preview }) => ({ url: src, base64: preview }))(criticalPlaceholder),
+        srcset: criticalImage.srcSet,
+        width: criticalImage.width,
+        height: criticalImage.height
+      }
+    }
+  }
+}
+</script>