+++ example/pages/lazy/font-extended/components/Lazy.vue
@@ -1,0 +1,31 @@
+<template>
+  <organism-preview-container>
+    <template #default>
+      <article
+        v-font="[
+          $fonts.getFont('Comic Neue', 400, 'normal'),
+          $fonts.getFont('Comic Neue', 700, 'normal', {selector: 'b, strong'}),
+          $fonts.getFont('Comic Neue', 400, 'italic', {selector: 'i, em'})
+        ]"
+        v-html="content"
+      />
+    </template>
+    <template #title>
+      <p>Font extended usage</p>
+    </template>
+  </organism-preview-container>
+</template>
+
+<script>
+import OrganismPreviewContainer from '@/components/organisms/PreviewContainer'
+export default {
+  components: {
+    OrganismPreviewContainer
+  },
+  data () {
+    return {
+      content: '<p>This is a <strong>RichText</strong> with different <i>font variants</i>.</p>'
+    }
+  }
+}
+</script>